Move aeson instances to cardano-api
