+++ cardano-api/cardano-api.cabal
@@ -123,6 +123,7 @@ library
                       , ouroboros-consensus-shelley
                       , ouroboros-network
                       , ouroboros-network-framework
+                      , plutus-core
                       , plutus-ledger-api
                       , scientific
                       , shelley-spec-ledger
